+++ core/dive.h
@@ -653,6 +653,8 @@ extern void vpmb_start_gradient(struct deco_state *ds);
 extern void vpmb_next_gradient(struct deco_state *ds, double deco_time, double surface_pressure);
 extern double tissue_tolerance_calc(struct deco_state *ds, const struct dive *dive, double pressure);
 extern bool is_dc_planner(const struct divecomputer *dc);
+extern bool has_planned(const struct dive *dive, bool planned);
+
 
 /* this should be converted to use our types */
 struct divedatapoint {